+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - EAS dev build config
 - Ability to vertically pan the image (not affecting slicing yet)
 - Auto bounce back when image panning is outside a nonsense range
+- Tap gesture placeholder in Image Slider
 
 ### Changed